image-to-text/app/requirements.txt
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
pytesseract==0.3.10
tesseract==0.1.3
